Major refactor, unfinished
