+++ pgbot/__init__.py
@@ -1,0 +1,7 @@
+from . import (
+    common,
+    util,
+    commands,
+    moderation,
+    sandbox
+)
